requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.0.9
annotated-types==0.6.0
anyio==4.3.0
anyjson==0.3.3
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
autopage==0.5.2
Babel==2.14.0
backend @ file:///home/ubuntu/cloudcomp-db/cloudcompa2
bcrypt==3.2.0
beautifulsoup4==4.10.0
bjoern==3.2.1
bleach==6.1.0
blinker==1.8.2
boto==2.49.0
branca==0.7.2
certifi==2021.10.8
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cloud-init==22.4.2
cmd2==2.4.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debtcollector==2.3.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.0.7
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
docopt==0.6.2
docutils==0.17.1
dogpile.cache==1.3.2
elastic-transport==8.13.0
elasticsearch==8.13.0
eventlet==0.30.2
exceptiongroup==1.2.1
executing==2.0.1
extras==1.0.0
fail2ban==0.11.2
fasteners==0.14.1
fastjsonschema==2.19.1
fixtures==3.0.0
Flask==3.0.3
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
gevent==21.12.0
gevent-ws==2.1.0
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
h11==0.14.0
heat-cfntools==1.4.2
hiredis==2.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
ipykernel==6.29.4
ipyleaflet==0.19.1
ipython==8.23.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
iso8601==1.0.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-leaflet==0.19.1
jupyter-lsp==2.2.5
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kafka-python==2.0.2
keyring==23.5.0
keystoneauth1==5.6.0
kiwisolver==1.4.5
kombu==5.2.3
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.5
munch==2.5.0
murano-agent==9.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.30.1
openstacksdk==3.1.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.0.1
oslo.concurrency==4.5.0
oslo.config==8.8.0
oslo.context==4.1.0
oslo.i18n==5.1.0
oslo.log==4.6.1
oslo.serialization==4.2.0
oslo.service==2.8.0
oslo.utils==4.12.2
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydocs==0.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-cinderclient==9.5.0
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-heatclient==3.5.0
python-json-logger==2.0.7
python-keystoneclient==5.4.0
python-ldap==3.2.0
python-magic==0.4.24
python-magnumclient==4.4.0
python-neutronclient==11.2.0
python-novaclient==18.6.0
python-octaviaclient==3.7.0
python-openstackclient==6.6.0
python-swiftclient==4.5.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.0.1
qtconsole==5.5.2
QtPy==2.4.1
redis==3.5.3
referencing==0.34.0
repoze.lru==0.7
requests==2.31.0
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
roman==3.3
Routes==2.5.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
semantic-version==2.8.5
Send2Trash==1.8.3
sentry-sdk==1.5.10
service-identity==18.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
sos==4.4
soupsieve==2.3.1
ssh-import-id==5.11
stack-data==0.6.3
stevedore==3.5.0
systemd-python==234
Tempita==0.5.2
tenacity==8.2.3
terminado==0.18.1
testtools==2.5.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traittypes==0.2.1
Twisted==22.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.7.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.13.3
xkit==0.0.0
XStatic==1.0.0
XStatic-jQuery==1.12.4.1
XStatic-JQuery-Migrate==1.2.1.1
XStatic-jquery-ui==1.12.0.1
XStatic-JQuery.Bootstrap.Wizard==1.0.0.1
XStatic-JQuery.quicksearch==2.0.4.1
XStatic-JQuery.TableSorter==2.14.5.1
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
